requirements.txt
asgiref==3.5.2
Django==4.0.8
djangorestframework==3.14.0
mssql-django==1.1.3
pyodbc==4.0.35
python-decouple==3.6
pytz==2022.6
sqlparse==0.4.3
